refactor Chat and useChat
