+++ packages/react/src/hooks/index.ts
@@ -43,3 +43,4 @@ export { UseTracksHookReturnType, UseTracksOptions, useTracks } from './useTrack
 export { UseVisualStableUpdateOptions, useVisualStableUpdate } from './useVisualStableUpdate';
 export { UseTrackOptions, useTrack } from './useTrack';
 export { useTrackByName } from './useTrackByName';
+export { useChat } from './useChat';
