+++ packages/react/src/prefabs/index.ts
@@ -1,12 +1,4 @@
-export { Chat, useChat, type ChatProps, type ChatMessage, type ReceivedChatMessage } from './Chat';
-export {
-  ChatEntry,
-  formatChatMessageLinks,
-  type ChatEntryProps,
-  type MessageDecoder,
-  type MessageEncoder,
-  type MessageFormatter,
-} from '../components/ChatEntry';
+export { Chat, type ChatProps } from './Chat';
 export {
   PreJoin,
   PreJoinProps,
@@ -18,3 +10,11 @@ export { VideoConference, type VideoConferenceProps } from './VideoConference';
 export { ControlBar, type ControlBarProps, type ControlBarControls } from './ControlBar';
 export { MediaDeviceMenu, type MediaDeviceMenuProps } from './MediaDeviceMenu';
 export { AudioConference, type AudioConferenceProps } from './AudioConference';
+
+// Re-export types from core
+export {
+  type ChatMessage,
+  type ReceivedChatMessage,
+  type MessageDecoder,
+  type MessageEncoder,
+} from '@livekit/components-core';
